+++ shared/hexabus_statemachine_structs.h
@@ -9,20 +9,20 @@
 // date/time transitions need to be stored separately. They are also executed seperately, each time before the "normal" transitions are executed
 
 struct condition {
-  uint8_t sourceIP[16];      // IP
-  uint8_t sourceEID;      // EID we expect data from
-  uint8_t op;           // predicate function						|
-  uint8_t datatype;			// The constant to compare with v
-	char 		data[4];			// Leave enough room for largest chunk of data, uint32_t/float in this case
+  uint8_t  sourceIP[16];    // IP
+  uint32_t sourceEID;       // EID we expect data from
+  uint8_t  op;              // predicate function           |
+  uint8_t  datatype;        // The constant to compare with v
+  char     data[4];         // Leave enough room for largest chunk of data, uint32_t/float in this case
 } __attribute__ ((packed));
 
 struct transition {
-  uint8_t fromState;      // current state
-  uint8_t cond;           // index of condition that must be matched
-  uint8_t eid;            // id of endpoint which should do something
-  uint8_t goodState;      // new state if everything went fine
-  uint8_t badState;       // new state if some went wrong
-  struct hxb_value value;  // Data for the endpoint
+  uint8_t  fromState;      // current state
+  uint8_t  cond;           // index of condition that must be matched
+  uint32_t eid;            // id of endpoint which should do something
+  uint8_t  goodState;      // new state if everything went fine
+  uint8_t  badState;       // new state if some went wrong
+  struct   hxb_value value;  // Data for the endpoint
 } __attribute__ ((packed));
 
 #define HXB_SM_HOUR    0x01
